break up big power converter; some general cleanup
