+++ examples/test_battery_protector.py
@@ -1,8 +1,6 @@
 import unittest
 
 from edg import *
-from electronics_lib.BatteryProtector_S8200A import BatteryProtector_S8200A
-from electronics_lib.DcDcConverters import Ltc3429
 
 
 class BatteryProtectorCircuit(SimpleBoardTop):
@@ -18,9 +16,6 @@ class BatteryProtectorCircuit(SimpleBoardTop):
     self.link_protect_neg = self.connect(self.battery_protector.gnd_out, self.led.gnd)
     self.link_protect_pos = self.connect(self.battery_protector.pwr_out, self.led.signal)
 
-
-if __name__ == "__main__":
-  compile_board_inplace(BatteryProtectorCircuit)
 
 class BatteryProtectorCircuitTestCase(unittest.TestCase):
   def test_design_battery_protector(self) -> None:
